+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat: Remove SentryOptions.enabled #736 !Breaking
 - fix: Setting environment for Sessions #734
 
 ## 6.0.0-beta.1